refactor: move entities to shared
