+++ src/pages/main-page/main-page.tsx
@@ -1,8 +1,8 @@
 import {FC, useCallback, useEffect} from 'react';
 import {useSearchParams} from 'react-router-dom';
-import {CITY_SEARCH_PARAM} from '../../entities/city/model/constants.ts';
-import {City} from '../../entities/city/model/types.ts';
-import {isValidCity} from '../../entities/city/model/utils.ts';
+import {CITY_SEARCH_PARAM} from '../../shared/entities/city/constants.ts';
+import {City} from '../../shared/entities/city/types.ts';
+import {isValidCity} from '../../shared/entities/city/utils.ts';
 import {CityLinkList} from '../../components/city-link-list/city-link-list.tsx';
 import {loadOffers, setCity} from '../../slices/offers-slice.ts';
 import {HeaderLogoLink} from '../../components/shared/header-logo-link.tsx';
