+++ src/pages/offer-page/feedback-form.tsx
@@ -1,6 +1,6 @@
 import {FC, FormEventHandler, Fragment, useEffect, useState} from 'react';
-import {isValidationError} from '../../entities/error/model/utils.ts';
-import {Offer} from '../../entities/offer/model/types.ts';
+import {isValidationError} from '../../shared/entities/error/utils.ts';
+import {Offer} from '../../shared/entities/offer/types.ts';
 import {addComment, handleCommentPostingResult, resetCommentPostingState} from '../../slices/offer-page-slice.ts';
 import {useAppDispatch, useAppSelector} from '../../shared/redux-helpers/typed-hooks.ts';
 import {RequestStatus} from '../../shared/server-interaction/request-status.ts';
