+++ src/shared/entities/city/constants.ts
