+++ src/shared/entities/error/types.ts
