+++ src/shared/entities/error/utils.ts
