+++ src/shared/entities/offer/constants.ts
