+++ src/shared/entities/offer/data-mappers.ts
