+++ src/shared/entities/user/types.ts
